Adjusted schema to better map business
Adjusted appointment association to connect with services entity -> every appointment is connected to a service provider, service, and client. Indirectly speaking, is also tied to business BUT due to the links, we don't have to connect them. 1-N due to how an appointment can have multiple services given but each service is tied to 1 appointment [99% sure, lmk if wrong]